requirements.txt
tensorflow>=2.20.0
opencv-python>=4.8.1.78
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
pytest>=7.4.3
python-multipart>=0.0.6
numpy>=1.24.3
pandas>=2.0.3
matplotlib>=3.7.2
seaborn>=0.13.0
python-dotenv>=1.0.0
pillow>=10.0.1
tqdm>=4.66.1
scipy>=1.10.0